# api/schemas/openapi.yaml
openapi: 3.0.3
info:
  title: LLM-CoPilot-Agent API
  version: 1.0.0
  description: |
    Comprehensive REST API for LLM-powered DevOps automation agent.

    **Features:**
    - Natural language to DevOps operations
    - Real-time streaming responses
    - Workflow orchestration
    - Observability queries
    - Incident management

    **Authentication:** JWT Bearer tokens required for all endpoints
    **Rate Limiting:** 1000 requests/hour per user, 100/minute burst
  contact:
    name: API Support
    email: api@llm-copilot-agent.dev
  license:
    name: Commercial License
    url: https://github.com/llm-copilot-agent/blob/main/LICENSE.md
servers:
  - url: https://api.llm-copilot-agent.dev/v1
    description: Production
  - url: https://staging-api.llm-copilot-agent.dev/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local Development
security:
  - bearerAuth: []
...
paths:
...
  /sessions:
    post:
      tags: [sessions]
      summary: Create new conversation session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags: [sessions]
      summary: List user sessions
      operationId: listSessions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, completed, expired]
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, last_activity]
            default: last_activity
      responses:
        '200':
          description: Sessions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sessions/{sessionId}:
    get:
      tags: [sessions]
      summary: Get session details
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: Session retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [sessions]
      summary: Delete session and all associated data
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '204':
          description: Session deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== MESSAGES ====================
  /sessions/{sessionId}/messages:
    post:
      tags: [messages]
      summary: Send message to session
      operationId: sendMessage
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Message processed successfully (non-streaming)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '202':
          description: Message accepted for async processing
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: URL to poll for message status
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [messages]
      summary: Get message history
      operationId: getMessageHistory
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: before
          in: query
          schema:
            type: string
            format: date-time
          description: Get messages before this timestamp
        - name: after
          in: query
          schema:
            type: string
            format: date-time
          description: Get messages after this timestamp
      responses:
        '200':
          description: Message history retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageHistoryResponse'
...
  /workflows:
    post:
      tags: [workflows]
      summary: Create and optionally execute workflow
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [workflows]
      summary: List workflows
      operationId: listWorkflows
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/WorkflowStatus'
        - name: template
          in: query
          schema:
            type: string
          description: Filter by workflow template name
      responses:
        '200':
          description: Workflows retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'
  /workflows/{workflowId}:
    get:
      tags: [workflows]
      summary: Get workflow status and details
      operationId: getWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
      responses:
        '200':
          description: Workflow retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /workflows/{workflowId}/execute:
    post:
      tags: [workflows]
      summary: Execute workflow
      operationId: executeWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteWorkflowRequest'
      responses:
        '202':
          description: Workflow execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token required. Include in Authorization header:
        `Authorization: Bearer <token>`
  parameters:
    SessionIdParam:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      description: Session UUID
    WorkflowIdParam:
      name: workflowId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Workflow UUID
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
  headers:
    X-RateLimit-Limit:
      schema:
        type: integer
      description: Total requests allowed per window
    X-RateLimit-Remaining:
      schema:
        type: integer
      description: Remaining requests in current window
    X-RateLimit-Reset:
      schema:
        type: integer
        format: unix-timestamp
      description: Unix timestamp when rate limit resets
  schemas:
...
    CreateSessionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[a-zA-Z0-9\s\-_]+$'
          description: Human-readable session name
          example: "Production deployment session"
        metadata:
          type: object
          additionalProperties:
            type: string
          maxProperties: 20
          description: Custom metadata (max 20 key-value pairs)
        preferences:
          $ref: '#/components/schemas/SessionPreferences'
        timeout_minutes:
          type: integer
          minimum: 5
          maximum: 1440
          default: 60
          description: Session timeout in minutes
    SessionPreferences:
      type: object
      properties:
        language:
          type: string
          enum: [en, es, fr, de, ja, zh]
          default: en
        timezone:
          type: string
          pattern: '^[A-Za-z]+/[A-Za-z_]+$'
          example: "America/New_York"
        stream_responses:
          type: boolean
          default: true
        include_context:
          type: boolean
          default: true
          description: Include conversation context in responses
    SessionResponse:
      type: object
      required: [id, user_id, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, paused, completed, expired]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_activity:
          type: string
          format: date-time
        message_count:
          type: integer
          minimum: 0
        metadata:
          type: object
          additionalProperties:
            type: string
        preferences:
          $ref: '#/components/schemas/SessionPreferences'
    SessionListResponse:
      type: object
      required: [sessions, pagination]
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    # ==================== MESSAGE SCHEMAS ====================
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 10000
          description: User message content
          example: "Show me CPU usage for auth-service"
        attachments:
          type: array
          maxItems: 10
          items:
            $ref: '#/components/schemas/MessageAttachment'
        context_override:
          type: object
          properties:
            include_history:
              type: boolean
              default: true
            max_history_turns:
              type: integer
              minimum: 0
              maximum: 50
              default: 10
    MessageAttachment:
      type: object
      required: [type, content]
      properties:
        type:
          type: string
          enum: [file, url, image, code]
        content:
          type: string
          maxLength: 100000
        filename:
          type: string
          maxLength: 255
        mime_type:
          type: string
          pattern: '^[a-z]+/[a-z0-9\-\+\.]+$'
    MessageResponse:
      type: object
      required: [id, session_id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        intent:
          $ref: '#/components/schemas/Intent'
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        created_at:
          type: string
          format: date-time
        metadata:
          type: object
          properties:
            model:
              type: string
            tokens_used:
              type: integer
            latency_ms:
              type: integer
            cache_hit:
              type: boolean
    MessageHistoryResponse:
      type: object
      required: [messages, pagination]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    # ==================== STREAMING SCHEMAS ====================
...
    CreateWorkflowRequest:
      type: object
      required: [name, tasks]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        template:
          type: string
          maxLength: 100
          description: Use predefined workflow template
        tasks:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/WorkflowTask'
        auto_execute:
          type: boolean
          default: false
        approval_required:
          type: boolean
          default: true
          description: Require human approval before execution
        rollback_on_failure:
          type: boolean
          default: false
...
    WorkflowStatus:
      type: string
      enum: [draft, pending_approval, approved, running, paused, completed, failed, cancelled, rolled_back]
    WorkflowResponse:
      type: object
      required: [id, name, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowTaskStatus'
        progress:
          type: object
          properties:
            total_tasks:
              type: integer
            completed_tasks:
              type: integer
            failed_tasks:
              type: integer
            progress_percentage:
              type: number
              format: float
              minimum: 0
              maximum: 100
...
    WorkflowListResponse:
      type: object
      required: [workflows, pagination]
      properties:
        workflows:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ExecuteWorkflowRequest:
      type: object
      properties:
        variables:
          type: object
          additionalProperties:
            type: string
          maxProperties: 50
        dry_run:
          type: boolean
          default: false
    WorkflowExecutionResponse:
      type: object
      required: [workflow_id, execution_id, status]
      properties:
        workflow_id:
          type: string
          format: uuid
        execution_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        started_at:
          type: string
          format: date-time

    # ==================== QUERY SCHEMAS ====================
...
    Intent:
      type: object
      required: [type, confidence]
      properties:
        type:
          type: string
          enum:
            - query_metrics
            - query_logs
            - query_traces
            - execute_workflow
            - create_incident
            - update_incident
            - deploy_service
            - scale_service
            - rollback_deployment
            - run_tests
            - analyze_performance
            - explain_error
            - general_question
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        parameters:
          type: object
          additionalProperties: true
    Entity:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          enum: [service, metric, log_level, time_range, environment, version, user, error_code]
        value:
          type: string
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        resolved_value:
          type: string
          description: Canonicalized/resolved entity value

    # ==================== COMMON SCHEMAS ====================
    PaginationInfo:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_previous:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        request_id:
          type: string
          format: uuid
          description: Unique request ID for tracing
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              field_errors:
                - field: "content"
                  message: "Content cannot be empty"
                  constraint: "minLength: 1"
    Unauthorized:
      description: Unauthorized - invalid or missing JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or expired authentication token"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Session not found"
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Rate limit exceeded. Please retry after 60 seconds"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
              request_id: "550e8400-e29b-41d4-a716-446655440000"